Move Locatable into the bag verifier
