+++ bag_verifier/src/main/scala/uk/ac/wellcome/platform/archive/bagverifier/storage/Locatable.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.common.storage
+package uk.ac.wellcome.platform.archive.bagverifier.storage
 
 import uk.ac.wellcome.storage.ObjectLocation
 
@@ -22,18 +22,3 @@ object Locatable {
   }
 }
 
-sealed trait LocateFailure[T] {
-  val t: T
-  val msg: String
-}
-
-// TODO: Check these error
-case class LocationError[T](t: T, msg: String)
-    extends Throwable(msg)
-    with LocateFailure[T]
-case class LocationNotFound[T](t: T, msg: String)
-    extends Throwable(msg)
-    with LocateFailure[T]
-case class LocationParsingError[T](t: T, msg: String)
-    extends Throwable(msg)
-    with LocateFailure[T]
